Moving metrics out into its own module under util
