+++ actors/stage/src/main/java/com/ea/orbit/actors/Stage.java
@@ -34,7 +34,8 @@ import com.ea.orbit.actors.cluster.ClusterPeer;
 import com.ea.orbit.actors.cluster.NodeAddress;
 import com.ea.orbit.actors.extensions.LifetimeExtension;
 import com.ea.orbit.actors.extensions.ActorExtension;
-import com.ea.orbit.actors.metrics.MetricsManager;
+import com.ea.orbit.metrics.MetricsManager;
+import com.ea.orbit.metrics.config.ReporterConfig;
 import com.ea.orbit.actors.runtime.Execution;
 import com.ea.orbit.actors.runtime.Hosting;
 import com.ea.orbit.actors.runtime.NodeCapabilities;
@@ -53,6 +54,7 @@ import org.slf4j.LoggerFactory;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
+import java.lang.reflect.Method;
 import java.time.Clock;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -76,11 +78,11 @@ public class Stage implements Startable
     @Config("orbit.actors.extensions")
     private List<ActorExtension> extensions = new ArrayList<>();
 
+    @Config("orbit.metrics.reporters")
+    private List<ReporterConfig> metricsConfig = new ArrayList<>();
+
     @Wired
     Container container;
-
-    @Inject
-    MetricsManager metricsManager;
 
     public enum StageMode
     {
@@ -246,12 +248,20 @@ public class Stage implements Startable
         hosting.start();
         execution.start();
 
-        if (metricsManager != null)
+        try
         {
+            Class mmClass = Class.forName("com.ea.orbit.metrics.MetricsManager");
+            Method factoryMethod = mmClass.getDeclaredMethod("getInstance");
+            factoryMethod.invoke(null, null);
+
             String cleanRuntimeId = runtimeIdentity().replace("Orbit", "");
             cleanRuntimeId = MetricsManager.sanitizeMetricName(getClusterName()) + "." + MetricsManager.sanitizeMetricName(getNodeName()) + "." + MetricsManager.sanitizeMetricName(cleanRuntimeId);
-            metricsManager.initializeMetrics(cleanRuntimeId);
-            metricsManager.registerExportedMetrics(execution);
+            MetricsManager.getInstance().initializeMetrics(cleanRuntimeId, metricsConfig);
+            MetricsManager.getInstance().registerExportedMetrics(execution);
+        }
+        catch(Exception ex)
+        {
+
         }
 
         Task<?> future = clusterPeer.join(clusterName, nodeName);
