+++ actors/stage/src/main/java/com/ea/orbit/actors/runtime/Execution.java
@@ -40,7 +40,7 @@ import com.ea.orbit.actors.extensions.InvokeHookExtension;
 import com.ea.orbit.actors.extensions.LifetimeExtension;
 import com.ea.orbit.actors.extensions.ActorExtension;
 import com.ea.orbit.actors.extensions.InvocationContext;
-import com.ea.orbit.actors.metrics.annotations.ExportMetric;
+import com.ea.orbit.metrics.annotations.ExportMetric;
 import com.ea.orbit.concurrent.ExecutorUtils;
 import com.ea.orbit.concurrent.Task;
 import com.ea.orbit.exception.UncheckedException;
