+++ samples/chat/chat-actors/src/main/resources/conf/orbit.yaml
@@ -41,6 +41,6 @@ orbit.components:
   - com.ea.orbit.actors.server.ServerModule
 
 orbit.metrics.reporters:
-  - !!com.ea.orbit.actors.metrics.config.reporters.Slf4jReporterConfig
+  - !!com.ea.orbit.metrics.config.Slf4jReporterConfig
     period: 1
     periodUnit: 'MINUTES'
